pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "news-aggregator"
version = "0.1.0"
description = "A simple news aggregator project"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "flask",
    "fastapi",
    "uvicorn",
    "psycopg2-binary",
    "python-dotenv",
    "jinja2",
    "pydantic",
    "requests",
    "spacy",
    "groq",
    "beautifulsoup4"
]

[project.scripts]
news-aggregator = "news_aggregator.cli:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["news_aggregator*"]
